Initial commit of Shader and Shader Manager code.
Shader allows you to load, compile and attach shaders.
ShaderManager caches loaded shaders and returns the cached object if you
request the same shader twice.
Soon, a Program class that uses Shaders to link programs that you can
bind, and that provides default Programs to be used by the rest of
aiObjectiveC.